Move the `futex` module back to being a submodule of `thread`.
This helps avoid adding more top-level module clutter, and futexes
are related to threads; they even have builtin knowledge of TIDs.
