+++ Cargo.toml
@@ -134,9 +134,6 @@ event = []
 # Enable `rustix::fs::*`.
 fs = []
 
-# Enable `rustix::futex::*`.
-futex = []
-
 # Enable `rustix::io_uring::*` (on platforms that support it).
 io_uring = ["event", "fs", "net", "linux-raw-sys/io_uring"]
 
@@ -147,7 +144,7 @@ mount = []
 net = ["linux-raw-sys/net", "linux-raw-sys/netlink", "linux-raw-sys/if_ether", "linux-raw-sys/xdp"]
 
 # Enable `rustix::thread::*`.
-thread = ["linux-raw-sys/prctl", "futex"]
+thread = ["linux-raw-sys/prctl"]
 
 # Enable `rustix::process::*`.
 process = ["linux-raw-sys/prctl"]
@@ -193,7 +190,6 @@ runtime = ["linux-raw-sys/prctl"]
 all-apis = [
     "event",
     "fs",
-    "futex",
     "io_uring",
     "mm",
     "mount",
