+++ src/backend/libc/thread/futex.rs
@@ -3,7 +3,7 @@ use crate::backend::c;
 bitflags::bitflags! {
     /// `FUTEX_*` flags for use with functions in [`futex`].
     ///
-    /// [`futex`]: crate::futex
+    /// [`futex`]: mod@crate::thread::futex
     #[repr(transparent)]
     #[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
     pub struct Flags: u32 {
@@ -16,7 +16,7 @@ bitflags::bitflags! {
 
 /// `FUTEX_*` operations for use with functions in [`futex`].
 ///
-/// [`futex`]: crate::futex
+/// [`futex`]: mod@crate::thread::futex
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 #[repr(u32)]
 pub enum Operation {
