+++ src/backend/libc/thread/mod.rs
@@ -1,2 +1,4 @@
+#[cfg(linux_kernel)]
+pub(crate) mod futex;
 #[cfg(not(windows))]
 pub(crate) mod syscalls;
