--- src/backend/linux_raw/futex/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod syscalls;
-pub(crate) mod types;
