+++ src/backend/linux_raw/thread/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod futex;
 pub(crate) mod syscalls;
