--- tests/futex/main.rs
@@ -1,6 +1,0 @@
-//! Tests for [`rustix::futex`].
-
-#![cfg(feature = "futex")]
-#![cfg(linux_kernel)]
-
-mod basic;
